+++ tests/apps/rename_and_drop_app/migrations/0002_rename_table_and_drop_column.py
@@ -1,0 +1,25 @@
+from django.db import migrations
+
+
+class Migration(migrations.Migration):
+
+    atomic = False
+
+    dependencies = [
+        ("rename_and_drop_app", "0001_initial"),
+    ]
+
+    operations = [
+        migrations.AlterModelTable(
+            name="testtable",
+            table="rename_and_drop_test_table_new",
+        ),
+        migrations.RemoveField(
+            model_name="testtable",
+            name="test_field_indexed",
+        ),
+        migrations.RemoveField(
+            model_name="testtable",
+            name="test_field_unique",
+        ),
+    ]